+++ include/susperia/internal/networking/http.h
@@ -5,14 +5,13 @@
 #ifndef SUSPIRIA_HTTP_H
 #define SUSPIRIA_HTTP_H
 
-#include <iostream>
-#include <sstream>
+#include <memory>
 #include <unordered_map>
-#include <functional>
-
-#include <mongoose/mongoose.h>
+#include <string>
 
 #include "routing/routing.h"
+#include "ip.h"
+#include "protocol.h"
 
 
 namespace suspiria {
@@ -25,6 +24,7 @@ namespace suspiria {
       BadRequest = 400,
     };
 
+
     class NewHttpRequest {
     public:
       std::string uri;
@@ -33,73 +33,18 @@ namespace suspiria {
       std::unordered_map<std::string, std::string> headers;
     };
 
-    class HttpRequest {
+
+    class http_protocol_factory : public protocol_factory {
     public:
-      explicit HttpRequest(RouteParams& params, mg_connection& connection, std::ostream& response_stream, std::istream& input_stream);
-      RouteParams& url_params;
-      std::istream& body;
-
-    private:
-      std::ostream& _response_stream;
-      mg_connection& _connection;
-
-      friend class StreamingResponse;
+      std::unique_ptr<protocol> create_protocol(tcp_connection& connection) override;
     };
 
-    class HttpResponse {
+
+    class http_server : public tcp_server {
     public:
-      explicit HttpResponse(HttpStatus status = HttpStatus::OK) : status(status) {}
-      std::unordered_map<std::string, std::string> headers;
-      HttpStatus status;
-
-      virtual void write(std::ostream& output);
-
-    protected:
-      void write_header(std::ostream& output);
-    };
-
-    class TextResponse : public HttpResponse {
-    public:
-      explicit TextResponse(std::string&& content, HttpStatus status = HttpStatus::OK);
-      void write(std::ostream& output) override;
-    private:
-      std::string _content;
-    };
-
-    class StreamingResponse : public HttpResponse {
-    public:
-      void prepare(HttpRequest& request, const std::function<void(std::ostream&)>& write_func);
-      void write(std::ostream& output) override { }
-    };
-
-    class HttpRequestHandler {
-    public:
-      virtual std::unique_ptr<HttpResponse> handle(HttpRequest& request) = 0;
-    };
-
-    class WebSocketServer {
-    public:
-
-      enum Status {
-        Listening,  // When actively listening.
-        Stopping,  // When the server is stopping all its operations and I/O loops. This is unlikely.
-        Idle,  // When the server is not doing anything. No I/O loop here.
-      };
-
-      explicit WebSocketServer(std::string address, std::shared_ptr<Router<HttpRequestHandler>> router);
-      ~WebSocketServer();
-      void start();
-      void stop();
-
-      const Status& get_status() const { return _status; }
-      int polling_timeout = 1000;  // timeout for every poll (in milliseconds)
-      inline const Router<HttpRequestHandler>& get_router() const { return *_router; }
-
-    private:
-      mg_mgr _manager;
-      std::string _address;
-      Status _status = Status::Idle;
-      std::shared_ptr<Router<HttpRequestHandler>> _router;
+      http_server(
+        asio::io_context& io, std::string host, unsigned short port
+      ) : tcp_server(io, std::move(host), port, std::make_shared<http_protocol_factory>()) {}
     };
 
   }
